+++ src/OpenGL/VertexBuffer.cpp
@@ -1,6 +1,6 @@
 #include "VertexBuffer.h"
 
-VertexBuffer::VertexBuffer(uint64_t size)
+VertexBuffer::VertexBuffer(uint32_t size)
 {
 	glCreateBuffers(1, &m_rendererID);
 	Bind();
@@ -22,7 +22,7 @@ void VertexBuffer::Unbind() const
 	glBindBuffer(GL_ARRAY_BUFFER, 0);
 }
 
-void VertexBuffer::StreamData(const void* data, uint64_t size, int64_t offset)
+void VertexBuffer::StreamData(const void* data, uint32_t size, int32_t offset)
 {
 	glBufferSubData(GL_ARRAY_BUFFER, offset, size, data);
 }
