+++ src/OpenGL/VertexBuffer.h
@@ -1,14 +1,15 @@
 #pragma once
 
+#include <cstdint>
+
 #include "glad/glad.h"
 
-#include <cstdint>
 
 class VertexBuffer
 {
 public:
 
-	VertexBuffer(uint64_t size);
+	VertexBuffer(uint32_t size);
 
 	~VertexBuffer();
 
@@ -16,7 +17,7 @@ public:
 
 	void Unbind() const;
 
-	void StreamData(const void* data, uint64_t size, int64_t offset);
+	void StreamData(const void* data, uint32_t size, int32_t offset);
 
 private:
 	unsigned int m_rendererID;
